Add missing docstring to qasm_utils
